third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/border-radius-expected.txt

This is a testharness.js-based test.
Found 32 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Can set 'border-top-left-radius' to a length: 0px
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-top-left-radius' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'border-top-left-radius' to a length: 3.14cm
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-top-left-radius' to a length: calc(0px + 0em)
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-top-left-radius' to a percent: 0%
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-top-left-radius' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'border-top-left-radius' to a percent: 3.14%
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-top-left-radius' to a percent: calc(0% + 0%)
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-top-right-radius' to a length: 0px
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-top-right-radius' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'border-top-right-radius' to a length: 3.14cm
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-top-right-radius' to a length: calc(0px + 0em)
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-top-right-radius' to a percent: 0%
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-top-right-radius' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'border-top-right-radius' to a percent: 3.14%
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-top-right-radius' to a percent: calc(0% + 0%)
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-bottom-left-radius' to a length: 0px
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-bottom-left-radius' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'border-bottom-left-radius' to a length: 3.14cm
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-bottom-left-radius' to a length: calc(0px + 0em)
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-bottom-left-radius' to a percent: 0%
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-bottom-left-radius' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'border-bottom-left-radius' to a percent: 3.14%
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-bottom-left-radius' to a percent: calc(0% + 0%)
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-bottom-right-radius' to a length: 0px
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-bottom-right-radius' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'border-bottom-right-radius' to a length: 3.14cm
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-bottom-right-radius' to a length: calc(0px + 0em)
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-bottom-right-radius' to a percent: 0%
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-bottom-right-radius' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'border-bottom-right-radius' to a percent: 3.14%
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
[FAIL] Can set 'border-bottom-right-radius' to a percent: calc(0% + 0%)
  assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
Harness: the test ran to completion.
